# docs/contracts/public-rest.openapi.yaml
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/auth/sessions": {
      "post": {
        "operationId": "Auth.login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "email",
                  "organizationSlug",
                  "devicePublicId",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "organizationSlug": {
                    "type": "string",
                    "maxLength": 80
                  },
                  "devicePublicId": {
                    "type": "string",
                    "maxLength": 128
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "writeOnly": true,
                    "maxLength": 1024
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSession"
                }
              }
            }
          },
          "default": {
            "description": "Stable non-sensitive error envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "summary": "Create an access and refresh session for an enrolled device",
        "tags": [
          "authentication"
        ],
        "x-swar-auth-kind": "PUBLIC",
        "x-swar-required-permissions": [],
        "x-swar-rate-limit-category": "SENSITIVE",
        "x-swar-idempotency": "NOT_APPLICABLE"
      }
    },
    "/api/v1/auth/sessions/refresh": {
      "post": {
        "operationId": "Auth.refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "refreshToken"
                ],
                "properties": {
                  "refreshToken": {
                    "type": "string",
                    "writeOnly": true,
                    "maxLength": 512
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSession"
                }
              }
            }
          },
          "default": {
            "description": "Stable non-sensitive error envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "summary": "Rotate a refresh session and issue fresh tokens",
        "tags": [
          "authentication"
        ],
        "x-swar-auth-kind": "PUBLIC",
        "x-swar-required-permissions": [],
        "x-swar-rate-limit-category": "SENSITIVE",
        "x-swar-idempotency": "NOT_APPLICABLE"
      }
    },
    "/api/v1/auth/sessions/revoke": {
      "post": {
        "operationId": "Auth.revoke",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "refreshToken"
                ],
                "properties": {
                  "refreshToken": {
                    "type": "string",
                    "writeOnly": true,
                    "maxLength": 512
                  }
                }
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          },
          "default": {
            "description": "Stable non-sensitive error envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "summary": "Revoke a refresh session",
        "tags": [
          "authentication"
        ],
        "x-swar-auth-kind": "PUBLIC",
        "x-swar-required-permissions": [],
        "x-swar-rate-limit-category": "SENSITIVE",
        "x-swar-idempotency": "NOT_APPLICABLE"
      }
    },
...
    "/api/v1/interventions/{interventionId}/verification-challenges": {
      "post": {
        "operationId": "Interventions.request",
        "parameters": [
          {
            "name": "interventionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVerificationChallengeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationChallenge"
                }
              }
            }
          },
          "default": {
            "description": "Stable non-sensitive error envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Request independent step-up or callback verification",
        "tags": [
          "interventions"
        ],
        "x-swar-auth-kind": "USER_ACCESS_JWT",
        "x-swar-required-permissions": [
          "intervention.resolve"
        ],
        "x-swar-rate-limit-category": "SENSITIVE",
        "x-swar-idempotency": "IDEMPOTENCY_KEY_REQUIRED"
      }
    },
...
  },
  "info": {
    "title": "SWAR REST and callback API",
    "description": "Versioned contracts for authenticated users, internal services, and signed LiveKit webhooks. Examples are fictional.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api/v1",
      "description": "Versioned SWAR API base path"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "mlService": {
        "type": "http",
        "scheme": "bearer"
      },
      "verifierService": {
        "type": "http",
        "scheme": "bearer"
      },
      "liveKitWebhook": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization"
      }
    },
    "schemas": {
...
      "CreateVerificationChallengeDto": {
        "type": "object",
        "properties": {
          "method": {
            "type": "string",
            "enum": [
              "CALLBACK",
              "OUT_OF_BAND"
            ]
          }
        },
        "required": [
          "method"
        ]
      },
...
      "ErrorEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "message",
          "requestId"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "AuthSession": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "accessToken",
          "refreshToken",
          "tokenType",
          "expiresIn",
          "principal"
        ],
        "properties": {
          "accessToken": {
            "type": "string",
            "readOnly": true,
            "x-swar-sensitive": true
          },
          "refreshToken": {
            "type": "string",
            "readOnly": true,
            "x-swar-sensitive": true
          },
          "tokenType": {
            "type": "string",
            "enum": [
              "Bearer"
            ]
          },
          "expiresIn": {
            "type": "integer",
            "minimum": 1
          },
          "principal": {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "userId",
              "membershipId",
              "organizationId",
              "deviceId",
              "roles"
            ],
            "properties": {
              "userId": {
                "type": "string",
                "format": "uuid"
              },
              "membershipId": {
                "type": "string",
                "format": "uuid"
              },
              "organizationId": {
                "type": "string",
                "format": "uuid"
              },
              "deviceId": {
                "type": "string",
                "format": "uuid"
              },
              "roles": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
      "VerificationChallenge": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "challengeId",
          "interventionId",
          "status",
          "method",
          "expiresAt"
        ],
        "properties": {
          "challengeId": {
            "type": "string",
            "format": "uuid"
          },
          "interventionId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "enum": [
              "PENDING",
              "PASSED",
              "FAILED",
              "EXPIRED",
              "CANCELLED"
            ]
          },
          "method": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "resultCode": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    }
  }
}